fix: keep Expo Go working
